+++ src/index.ts
@@ -1,6 +1,5 @@
 import { Viewer } from "./Viewer";
 import { Configuration } from "./Configuration/Configuration";
-import axios from "axios";
 import { Excursion } from "./Models/ExcursionModels/Excursion";
 import { BuildConfiguration } from "./Configuration/BuildConfiguration";
 
@@ -9,27 +8,29 @@ const supportedTourVersion = "v0.9";
 document.addEventListener("DOMContentLoaded", async () => {
 
     console.log(`excursion viewer ${supportedTourVersion}`);
-    const configuration = await axios.get<Configuration>(BuildConfiguration.ConfigFilePath || "config.json");
-    if (configuration.status !== 200) {
+    const configurationResponse = await fetch(BuildConfiguration.ConfigFilePath || "config.json");
+    if (configurationResponse.status !== 200) {
         console.warn("Can't get configuration");
         return;
     }
-    const viewer = new Viewer(configuration.data);
+    const configuration = await (configurationResponse).json() as Configuration;
+    const viewer = new Viewer(configuration);
     viewer.createScene();
     (document as any).viewer = viewer;
     try {
-        const response = await axios.get<Excursion>(configuration.data.sceneUrl + "tour.json");
-        if (response.status !== 200) {
+        const tourResonses = await fetch(configuration.sceneUrl + "tour.json");
+        if (tourResonses.status !== 200) {
             console.warn("Can't get scene description");
             return;
         }
-        const excursion = response.data;
+        const excursion = await (await fetch(configuration.sceneUrl + "tour.json")).json() as Excursion;
+
         if (!excursion.tourProtocolVersion) {
             alert("Too old protocol (without version), use new builder or old viewer");
             return;
         }
         if (excursion.tourProtocolVersion != supportedTourVersion) {
-            alert(`That viewer supports only tour ${supportedTourVersion}, please use another viewer or builder (now try ${response.data.tourProtocolVersion})`);
+            alert(`That viewer supports only tour ${supportedTourVersion}, please use another viewer or builder (now try ${excursion.tourProtocolVersion})`);
             return;
         }
         console.table({
@@ -42,6 +43,6 @@ document.addEventListener("DOMContentLoaded", async () => {
         await viewer.show(excursion);
     } catch (error) {
         console.error(error);
-        alert(`Can't load excursion from ${configuration.data.sceneUrl}`);
+        alert(`Can't load excursion from ${configuration.sceneUrl}`);
     }
 });